backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
chromadb==0.5.3
sentence-transformers==3.0.1
langchain-text-splitters==0.2.2
pydantic==2.7.4
python-dotenv==1.0.1
python-docx==1.1.2
httpx==0.27.0
requests==2.32.3
openai==1.35.0